requirements.txt
flask
flask-cors
python-dotenv
google-genai
pydantic>=2.7
gunicorn
flask
